pyproject.toml
[project]
name = "bc-z"
version = "0.1.0"
description = "Educational BC-Z implementation - Zero-Shot Task Generalization with Imitation Learning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "h5py>=3.8.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "sentence-transformers>=2.2.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "ipython>=9.8.0",
    "tensorflow>=2.20.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "wandb>=0.15.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
